Reintroduced tabular layout for settings page
