+++ assets/templates/account/settings/settings_authorisation.html
@@ -1,0 +1,11 @@
+{% extends 'account/settings/settings_base.html' %}
+
+{% block tab_authorisation %}active{% endblock %}
+
+
+{% block details %}
+
+    See password_change.html for code
+    This is in a seperate file for allauth
+
+{% endblock details %}
