+++ assets/templates/account/settings/settings_mails.html
@@ -1,0 +1,11 @@
+{% extends 'account/settings/settings_base.html' %}
+
+{% block tab_mail %}active{% endblock %}
+
+
+{% block details %}
+
+    See email.html for code
+    This is in a seperate file for allauth
+
+{% endblock details %}
